Add optional authentication functionality (#29)
* Add optional authentication functionality

If you have private VMs you have to provide a way to authenticate to the
container registry in order to pull down the container. This adds a new
optional auth block with a username and password property.

If this block is set, and is valid, we will issue the tart login command
before we perform the clone operation.

* Add some state tests